refactor(Observatoire): déplacer le code backend 'statistics' dans un dossier dédié
