+++ api/views/__init__.py
@@ -64,7 +64,6 @@ from .reservationexpe import ReservationExpeView  # noqa: F401
 from .resourceaction import ResourceActionView  # noqa: F401
 from .review import ReviewView  # noqa: F401
 from .sector import SectorListView  # noqa: F401
-from .statistics import CanteenStatisticsView  # noqa: F401
 from .subscription import SubscribeNewsletter  # noqa: F401
 from .teledeclaration import (  # noqa: F401
     TeledeclarationCampaignDatesListView,
