+++ macantine/settings.py
@@ -88,14 +88,15 @@ INSTALLED_APPS = WAGTAIL_INSTALLED_APPS + [
     "ckeditor",
     "ckeditor_uploader",
     "macantine",
+    "common",
     "data",
     "api",
+    "statistics",
     "web",
     "magicauth",
     "django_extensions",
     "django_filters",
     "django_celery_results",
-    "common",
     "drf_spectacular",
     "drf_spectacular_sidecar",
     "simple_history",
