+++ statistics/README.md
@@ -1,0 +1,13 @@
+# Application `statistics`
+
+## Objectif
+
+Cette application regroupe la logique de lecture et de calcul des statistiques exposees par l'API.
+
+## URL publique
+
+Endpoint public :
+
+- `/api/v1/canteenStatistics/`
+
+Le routage est delegue par `api/urls.py` vers `statistics/urls.py`.
