+++ statistics/__init__.py
@@ -1,0 +1,1 @@
+"""Statistics Django app."""
