+++ statistics/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from statistics.views import CanteenStatisticsView
+
+urlpatterns = [
+    path("canteenStatistics/", CanteenStatisticsView.as_view(), name="canteen_statistics"),
+]
